+++ Parameters.h
@@ -34,7 +34,7 @@ struct Vehicle {
 };
 
 struct Parameters {
-    std::uint32_t theTimeSlotsCount;
+    std::size_t theTimeSlotsCount;
 
     std::vector<Front> theFronts;
     std::vector<Vehicle> theVehicles;
